+++ lending/CodatLending/Models/Operations/CreateConnectionRequestBody.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Operations
+namespace Codat.Lending.Models.Operations
 {
     using Newtonsoft.Json;