+++ lending/CodatLending/Models/Operations/GenerateLoanSummaryResponse.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Operations
+namespace Codat.Lending.Models.Operations
 {
-    using CodatLending.Models.Shared;
+    using Codat.Lending.Models.Shared;
     using System.Net.Http;
     using System;
     
@@ -35,6 +35,6 @@ namespace CodatLending.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }