+++ lending/CodatLending/Models/Operations/GetDataStatusResponse.cs
@@ -8,9 +8,9 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Operations
+namespace Codat.Lending.Models.Operations
 {
-    using CodatLending.Models.Shared;
+    using Codat.Lending.Models.Shared;
     using System.Collections.Generic;
     using System.Net.Http;
     using System;
@@ -26,7 +26,7 @@ namespace CodatLending.Models.Operations
         /// <summary>
         /// OK
         /// </summary>
-        public Dictionary<string, DataStatus>? DataStatusResponse { get; set; }
+        public Dictionary<string, DataStatus>? DataStatuses { get; set; }
 
         /// <summary>
         /// Your API request was not properly authorized.
@@ -41,6 +41,6 @@ namespace CodatLending.Models.Operations
         /// <summary>
         /// Raw HTTP response; suitable for custom response parsing
         /// </summary>
-        public HttpResponseMessage? RawResponse { get; set; }
+        public HttpResponseMessage RawResponse { get; set; } = default!;
     }
 }