+++ lending/CodatLending/Models/Shared/AccountingTransfer.cs
@@ -8,7 +8,7 @@
 // </auto-generated>
 //------------------------------------------------------------------------------
 #nullable enable
-namespace CodatLending.Models.Shared
+namespace Codat.Lending.Models.Shared
 {
     using Newtonsoft.Json;
     using System.Collections.Generic;
@@ -25,10 +25,10 @@ namespace CodatLending.Models.Shared
     {
 
         /// <summary>
-        /// The customer or supplier for the transfer, if available.
+        /// A customer or supplier associated with the direct cost.
         /// </summary>
         [JsonProperty("contactRef")]
-        public AccountingTransferContactRef? ContactRef { get; set; }
+        public ContactRef? ContactRef { get; set; }
 
         /// <summary>
         /// In Codat&apos;s data model, dates and times are represented using the &lt;a class=&quot;external&quot; href=&quot;https://en.wikipedia.org/wiki/ISO_8601&quot; target=&quot;_blank&quot;&gt;ISO 8601 standard&lt;/a&gt;. Date and time fields are formatted as strings; for example:<br/>