+++ lending/RELEASES.md
@@ -153,3 +153,13 @@ Based on:
 - [csharp v4.3.0] lending
 ### Releases
 - [NuGet v4.3.0] https://www.nuget.org/packages/Codat.Lending/4.3.0 - lending
+
+## 2023-12-05 11:20:18
+### Changes
+Based on:
+- OpenAPI Doc 3.0.0 https://raw.githubusercontent.com/codatio/oas/main/yaml/Codat-Lending.yaml
+- Speakeasy CLI 1.125.1 (2.210.3) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [csharp v5.0.0] lending
+### Releases
+- [NuGet v5.0.0] https://www.nuget.org/packages/Codat.Lending/5.0.0 - lending